+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New
 
 - implement `impute_fixed`, `impute_zero`, and `impute_factor`. notably these do not implement "scoped variants" which were previously implemented - for example, `impute_fixed_if` etc. This is in favour of using the new `across` workflow within `dplyr`, and it is easier to maintain. #261
+- add `digit` argument to `miss_var_summary` to help display %missing data correctly when there is a very small fraction of missingness. #284
 
 # naniar 1.0.0